Create new files for classes and fix edit text function
