+++ src/TaskClass.js
@@ -1,0 +1,7 @@
+export default class Task {
+  constructor(index, description, completed = false) {
+    this.index = index;
+    this.description = description;
+    this.completed = completed;
+  }
+}
